requirements.txt
streamlit
pandas
PyPDF2
google-generativeai
openpyxl
